+++ CHANGELOG.md
@@ -7,3 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.0] - 2019-07-27
 ### Added
 - Initial release of the tool.
+
+## [1.1.0] - 2019-07-27
+### Added
+- Support for Unity's shortcut manager
+### Changed
+- Upgrade from Unity 2018.3 to 2019.1
+- Replaced deprecated scene view GUI callback
+- Replaced deprecated preference item with new settings provider